+++ idea/testData/refactoring/pullUp/k2k/privateMemberWithUsagesToInterface.kt
@@ -1,0 +1,8 @@
+interface I
+
+class <caret>Foo : I {
+    // INFO: {checked: "true"}
+    private fun privateFun() = 0
+
+    fun refer() = privateFun()
+}